requirements.txt
# ============================================================================
# gemma4_comp - starter dependencies
# ----------------------------------------------------------------------------
# Pin exact versions once the training pipeline is stable. For now these are
# the upper bounds that are known to be compatible with Gemma 4 + Unsloth as
# of April 2026. Install with: pip install -r requirements.txt
# ============================================================================

# Fine-tuning (Special Tech track)
unsloth>=2025.10
transformers>=4.50.0
accelerate>=1.2.0
peft>=0.15.0
bitsandbytes>=0.46.0
datasets>=3.0.0
trl>=0.15.0
sentencepiece>=0.2.0

# Gemma / tokenizer
tokenizers>=0.21.0

# Deployment (Special Tech tracks)
# llama-cpp-python provides the GGUF runtime for the desktop demo.
llama-cpp-python>=0.3.0
# LiteRT (TFLite successor) - uncomment when ready to produce the mobile artifact.
# ai-edge-litert>=1.0.0

# Evaluation / metrics
numpy>=2.0.0
pandas>=2.2.0
scikit-learn>=1.5.0
scipy>=1.14.0

# Data models (matches existing framework)
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Config / YAML parsing
pyyaml>=6.0.0

# Web demo
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.20

# Persistence
sqlalchemy>=2.0.0

# Dev / tooling
pytest>=8.3.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
mypy>=1.13.0
